Fix int */sqInt * mismatch for Win64 FilePlugin
While at it, also revise a few potential mismatch for legacy Mac OS (which did not have a 64 bits flavour, but well...)

+++ platforms/Mac OS/vm/sqMacNSPlugin.c
@@ -766,7 +766,7 @@ primitivePluginRequestFileHandle(void) {
 	   folder will be--it's folder might even be user-name dependent. sqFileOpen() will
 	   only allow opening files within the sandbox, which would be better. Here's the
 	   sandbox-safe alternative:
-		  sqFileOpen(filePtr, (int) fileName, strlen(fileName), false);
+		  sqFileOpen(filePtr, fileName, strlen(fileName), false);
 	*/
 	OpenFileReadOnly(filePtr, fileName);
